# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Products Catalog API
  version: 1.0.0
paths:
  /tenants/{tenantId}/search/products:
    get:
      summary: Search products
      tags:
        - Products
      parameters:
        - in: path
          name: tenantId
          schema:
            type: string
          required: true
          description: Tenant identifier
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Search query
        - in: query
          name: lang
          schema:
            type: string
            enum: [ca, en, es, fr, gl, pt]
            default: es
          required: false
          description: Language code to search in (defaults to 'es')
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: Page number for pagination (0-based)
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 12
          required: false
          description: Results per page
        - in: query
          name: supplierId
          schema:
            type: string
          required: false
          description: Supplier id
        - in: query
          name: ref
          schema:
            type: string
          required: false
          description: Comma separated product refs
          example: "462"
        - in: query
          name: ean
          schema:
            type: string
          required: false
          description: Comma separated product eans
          example: "3253561801174"
        - in: query
          name: l1Id
          schema:
            type: string
          required: false
          description: Level1 cat id
          example: "06"
        - in: query
          name: l2Id
          schema:
            type: string
          required: false
          description: Level2 cat id
          example: "0604"
        - in: query
          name: l3Id
          schema:
            type: string
          required: false
          description: Level3 cat id
          example: "060406"
        - in: query
          name: id
          schema:
            type: string
          required: false
          description: Comma separated product ids
          example: "060406,AS234"
        - in: query
          name: visibility
          schema:
            type: integer
            default: 0
          required: false
          description: Client visibility
        - in: query
          name: sortBy
          schema:
            type: string
          required: false
          description: Sort by field; suffix with `+`/`-` for direction (e.g. `price+`)
          example: "price+"
        - in: query
          name: type
          schema:
            type: string
          required: false
          description: Product type
          example: "own"
        - in: query
          name: grouping
          schema:
            type: boolean
            default: false
          required: false
          description: Group products by grouping_code
        - in: query
          name: rate
          schema:
            type: string
          required: false
          description: Product rate id
          example: "R1"
      responses:
        "200":
          description: Matching products (full document in selected language)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSearchResponse"
  /tenants/{tenantId}/search/categories:
    get:
      summary: For a specific search, returns categories with each category total count. Used in category filters.
      description: Returns the level1/level2/level3 category hierarchy for the given search, with a product count per level3 category.
      tags:
        - Products
      parameters:
        - in: path
          name: tenantId
          schema:
            type: string
          required: true
          description: Tenant identifier
        - in: query
          name: q
          schema:
            type: string
          required: false
          description: Search query
        - in: query
          name: lang
          schema:
            type: string
            enum: [ca, en, es, fr, gl, pt]
            default: es
          required: false
          description: Language code to search in (defaults to 'es')
        - in: query
          name: supplierId
          schema:
            type: string
          required: false
          description: Supplier id
        - in: query
          name: ref
          schema:
            type: string
          required: false
          description: Comma separated product refs
          example: "462"
        - in: query
          name: ean
          schema:
            type: string
          required: false
          description: Comma separated product eans
          example: "3253561801174"
        - in: query
          name: l1Id
          schema:
            type: string
          required: false
          description: Level1 cat id
          example: "06"
        - in: query
          name: l2Id
          schema:
            type: string
          required: false
          description: Level2 cat id
          example: "0604"
        - in: query
          name: l3Id
          schema:
            type: string
          required: false
          description: Level3 cat id
          example: "060406"
        - in: query
          name: id
          schema:
            type: string
          required: false
          description: Comma separated product ids
          example: "060406,AS234"
        - in: query
          name: visibility
          schema:
            type: integer
            default: 0
          required: false
          description: Client visibility
        - in: query
          name: type
          schema:
            type: string
          required: false
          description: Product type
          example: "own"
        - in: query
          name: rate
          schema:
            type: string
          required: false
          description: Product rate id
          example: "R1"
      responses:
        "200":
          description: Array of categories (with product count) matching the search
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchCategoriesResponse"
  /tenants/{tenantId}/categories-tree:
    get:
      summary: Get categories tree for suppliers
      description: Returns a hierarchical tree of categories (level1 → level2 → level3) for the given supplier IDs, localized by the requested language.
      tags:
        - Categories
      parameters:
        - in: path
          name: tenantId
          schema:
            type: string
          required: true
          description: Tenant identifier
        - in: query
          name: lang
          schema:
            type: string
            enum: [ca, en, es, fr, gl, pt]
            default: es
          required: false
          description: Language code
        - in: query
          name: supplierIds
          schema:
            type: string
          required: false
          description: Comma separated supplier ids
          example: "WBG,BEN"
        - in: query
          name: visibility
          schema:
            type: integer
            default: 0
          required: false
          description: Client visibility
        - in: query
          name: deleted
          schema:
            type: boolean
          required: false
          description: Include deleted products. If not set, only non-deleted are returned.
      responses:
        "200":
          description: Hierarchical categories tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCategoriesTreeResponse"
components:
  schemas:
    CategoryNode:
      type: object
      properties:
        id:
          type: string
          description: Category identifier (level1, level2, or level3)
          example: "level1-cat"
        description:
          type: string
          description: Localized category description
          example: "Categoría Nivel 1"
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategoryNode"
    GetCategoriesTreeResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupplierCategoriesTree"
    CategoryCount:
      type: object
      properties:
        level1:
          type: string
        level1Name:
          type: string
        level2:
          type: string
        level2Name:
          type: string
        level3:
          type: string
        level3Name:
          type: string
        count:
          type: integer
    SearchCategoriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
    ProductSearchResult:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
        ean:
          type: string
        containerUnits:
          type: string
        containerType:
          type: string
        supplierId:
          type: string
        supplierName:
          type: string
        step:
          type: integer
          example: 1
        description:
          type: string
        shortDescription:
          type: string
        mainPictureUrl:
          type: string
        mainPictureThumbUrl:
          type: string
        level1:
          type: string
        level2:
          type: string
        level3:
          type: string
        level1Name:
          type: string
        level2Name:
          type: string
        level3Name:
          type: string
        discount1:
          type: number
        discount2:
          type: number
        netPrice:
          type: number
        salePrice:
          type: number
        netPriceWithMargin:
          type: number
        vatAmount:
          type: number
        isReplacementOf:
          type: boolean
          description: True when the search term exactly matched one of this product's deprecated `replaces` references.
        groupedItems:
          type: array
          description: When grouping is enabled, other products collapsed under the same grouping_code.
          items:
            $ref: "#/components/schemas/ProductSearchResult"
    ProductSearchResponse:
      type: object
      properties:
        navigation:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            count:
              type: integer
            limit:
              type: integer
            firstPage:
              type: string
              format: uri
            lastPage:
              type: string
              format: uri
            previousPage:
              type: string
              format: uri
              nullable: true
            nextPage:
              type: string
              format: uri
              nullable: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductSearchResult"
    SupplierCategoriesTree:
      type: object
      properties:
        supplierId:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryNode"